% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_with_dir.R
\name{make_with_dir}
\alias{make_with_dir}
\title{Create a pipeline using roxygen tags}
\usage{
make_with_dir(
  dir = ".",
  recursive = FALSE,
  build = TRUE,
  quiet = getOption("makepipe.quiet")
)
}
\arguments{
\item{dir}{A character vector of full path names; the default corresponds to the working directory}

\item{recursive}{A logical determining whether or not to recurse into
subdirectories}

\item{build}{A logical determining whether or not the pipeline will be built
immediately or simply returned to the user}

\item{quiet}{A logical determining whether or not messages are signaled}
}
\value{
A \code{Pipeline} object
}
\description{
Create a pipeline using roxygen tags
}
\examples{
\dontrun{
# Create a pipeline from scripts in the working dir without executing it
p <- make_with_dir(build = FALSE)
p$build() # Then execute it yourself
}
}
\seealso{
Other make: 
\code{\link{make_with_recipe}()},
\code{\link{make_with_source}()}
}
\concept{make}
