% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_with_recipe}
\alias{make_with_recipe}
\title{Make targets out of dependencies using a recipe}
\usage{
make_with_recipe(
  recipe,
  targets,
  dependencies,
  envir = parent.frame(),
  quiet = getOption("piper.quiet"),
  ...
)
}
\arguments{
\item{recipe}{A chunk of R code which makes the \code{targets}}

\item{targets}{A character vector of paths to files}

\item{dependencies}{A character vector of paths to files which the \code{targets}
depend on}

\item{envir}{The environment in which to run \code{recipe}}

\item{quiet}{A logical determining whether or not messages are signaled}

\item{...}{Additional parameters to pass to \code{base::eval()}}
}
\value{
The result of evaluating the \code{recipe} if the \code{targets} are out of
date otherwise `NULL``
}
\description{
Make targets out of dependencies using a recipe
}
\examples{

\dontrun{
# Merge files in fresh environment if raw data has been updated
# since last merged
make_with_recipe(
  recipe = {
    dat <- readRDS("data/raw_data.Rds")
    pop <- readRDS("data/pop_data.Rds")
    merged_dat <- merge(dat, pop, by = "id")
    saveRDS(merged_dat, "data/merged_data.Rds")
  },
  targets = "data/merged_data.Rds",
  dependencies = c("data/raw_data.Rds", "data/raw_pop.Rds"),
  envir = new.env()
)
}
}
\seealso{
Other make: 
\code{\link{make_with_source}()}
}
\concept{make}
