% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment.R
\name{SegmentRecipe}
\alias{SegmentRecipe}
\title{Segment}
\description{
A Segment object is automatically constructed and attached to
the Pipeline when a call to \verb{make_*()} is made. It stores the relationships
between targets, dependencies, and sources.
}
\seealso{
Other segment: 
\code{\link{SegmentSource}},
\code{\link{Segment}}
}
\concept{segment}
\keyword{internal}
\section{Super class}{
\code{\link[makepipe:Segment]{makepipe::Segment}} -> \code{SegmentRecipe}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recipe}}{A chunk of R code which makes the \code{targets}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{edges}}{Construct edges connecting the dependencies, instructions, and targets}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SegmentRecipe$new()}}
\item \href{#method-print}{\code{SegmentRecipe$print()}}
\item \href{#method-update_result}{\code{SegmentRecipe$update_result()}}
\item \href{#method-execute}{\code{SegmentRecipe$execute()}}
\item \href{#method-clone}{\code{SegmentRecipe$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new Segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentRecipe$new(
  id,
  recipe,
  targets,
  dependencies,
  packages,
  envir,
  force,
  executed,
  result,
  execution_time
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An integer that uniquely identifies the segment}

\item{\code{recipe}}{A chunk of R code which makes the \code{targets}}

\item{\code{targets}}{A character vector of paths to files}

\item{\code{dependencies}}{A character vector of paths to files which the \code{targets}
depend on}

\item{\code{packages}}{A character vector of names of packages which \code{targets}
depend on}

\item{\code{envir}}{The environment in which to execute the instructions.}

\item{\code{force}}{A logical determining whether or not execution of the \code{source}
or \code{recipe} will be forced (i.e. happen whether or not the targets are
out-of-date)}

\item{\code{executed}}{A logical, whether or not the instructions were executed}

\item{\code{result}}{An object, whatever is returned by executing the instructions}

\item{\code{execution_time}}{A difftime, the time taken to execute the instructions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printing method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentRecipe$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_result"></a>}}
\if{latex}{\out{\hypertarget{method-update_result}{}}}
\subsection{Method \code{update_result()}}{
Update the Segment with new execution information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentRecipe$update_result(executed, execution_time, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{executed}}{A logical, whether or not the instructions were executed}

\item{\code{execution_time}}{A difftime, the time taken to execute the instructions}

\item{\code{result}}{An object, whatever is returned by executing the instructions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Execute the Segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentRecipe$execute(envir = NULL, quiet = getOption("makepipe.quiet"), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{envir}}{The environment in which to execute the \code{source} or \code{recipe}. By
default, execution will take place in a fresh environment whose parent is
the calling environment.}

\item{\code{quiet}}{A logical determining whether or not messages are signaled}

\item{\code{...}}{Additional parameters to pass to \code{base::eval()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SegmentRecipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
