% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{make_params}
\alias{make_params}
\title{Parameters for make-like functions}
\arguments{
\item{targets}{A character vector of paths to files}

\item{dependencies}{A character vector of paths to files which the \code{targets}
depend on}

\item{packages}{A character vector of names of packages which \code{targets}
depend on}

\item{quiet}{A logical determining whether or not messages are signaled}
}
\description{
Parameters for make-like functions
}
\keyword{internal}
