% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_with_source}
\alias{make_with_source}
\title{Make targets out of dependencies using a source file}
\usage{
make_with_source(
  source,
  targets,
  dependencies,
  packages = NULL,
  quiet = getOption("makepipe.quiet"),
  ...
)
}
\arguments{
\item{source}{The path to an R script which makes the \code{targets}}

\item{targets}{A character vector of paths to files}

\item{dependencies}{A character vector of paths to files which the \code{targets}
depend on}

\item{packages}{A character vector of names of packages which \code{targets}
depend on}

\item{quiet}{A logical determining whether or not messages are signaled}

\item{...}{Additional parameters to pass to \code{base::source()}}
}
\value{
\code{NULL} invisibly.
}
\description{
Make targets out of dependencies using a source file
}
\examples{
\dontrun{
# Merge files in fresh environment if raw data has been updated
# since last merged
make_with_source(
  source = "merge_data.R",
  targets = "data/merged_data.Rds",
  dependencies = c("data/raw_data.Rds", "data/raw_pop.Rds"),
  local = new.env()
)
}

}
\seealso{
Other make: 
\code{\link{make_with_recipe}()}
}
\concept{make}
