% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_with_source}
\alias{make_with_source}
\title{Make targets out of dependencies using a source file}
\usage{
make_with_source(source, targets, dependencies, ...)
}
\arguments{
\item{source}{The path to an R script which makes the \code{targets}}

\item{targets}{A character vector of paths to files}

\item{dependencies}{A character vector of paths to files which the \code{targets} depend on}

\item{...}{Additional parameters to pass to \code{base::source()}}
}
\value{
\code{NULL} invisibly.
}
\description{
Make targets out of dependencies using a source file
}
\examples{

\dontrun{
# TODO
}

}
