% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{pipeline-accessors}
\alias{pipeline-accessors}
\alias{set_pipeline}
\alias{get_pipeline}
\alias{show_pipeline}
\alias{is_pipeline}
\title{Access Pipeline.}
\usage{
set_pipeline(pipeline)

get_pipeline()

show_pipeline(pipeline = get_pipeline())

is_pipeline(pipeline)
}
\arguments{
\item{pipeline}{A pipeline. See \link{Pipeline} for more details.}
}
\description{
\code{get_pipeline()} and \code{set_pipeline()} access and modify the current "active"
pipeline.
}
\keyword{internal}
