% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{Pipeline}
\alias{Pipeline}
\title{Pipeline visualisations}
\description{
A Pipeline object is automatically constructed as calls to \verb{make_*()} are
made. It stores the relationships between targets, dependencies, and
sources.
}
\examples{
\dontrun{
# TODO
}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{edges}}{A data frame}

\item{\code{nodes}}{A data frame}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pipeline_network}{\code{Pipeline$pipeline_network()}}
\item \href{#method-refresh_nodes}{\code{Pipeline$refresh_nodes()}}
\item \href{#method-print}{\code{Pipeline$print()}}
\item \href{#method-clone}{\code{Pipeline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pipeline_network"></a>}}
\if{latex}{\out{\hypertarget{method-pipeline_network}{}}}
\subsection{Method \code{pipeline_network()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pipeline$pipeline_network(dependencies, source, targets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dependencies}}{A character vector of paths to files which the \code{targets} depend on}

\item{\code{source}}{The path to an R script which makes the \code{targets}}

\item{\code{targets}}{A character vector of paths to files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-refresh_nodes}{}}}
\subsection{Method \code{refresh_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pipeline$refresh_nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pipeline$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
